requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
gunicorn==20.1.0
huggingface-hub==0.16.4
nltk==3.8.1
numpy==1.25.2
openai==0.27.8
pinecone-client==2.2.2
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
tiktoken==0.4.0
tokenizers==0.13.3
transformers==4.31.0
urllib3==2.0.4